Block initialization, not subclassing on missing PACKET_ID
The Packet class was originally using __init_subclass__, which is called
each time a child class is created. In there, we were performing a check
that ensures that this child class does define PACKET_ID class variable.

This does works, however it is against the nature of ABCs, since it
doesn't allow the class to be extendable through another abstract class,
it forces all subclasses to already have PACKET_ID, which is needed for
concrete subclasses, but not necessarily for abstract subclasses, hence
preventing more abstract subclasses to be made from this class.

Instead of that, this commit replaces __init_subclass__ with simple
check in __init__, as every subclass is expected to call the __init__ of
it's super class, triggering this PACKET_ID check with it. Since
__init__ only runs on initialization (making of an instance), not on
making of another class itself, it behaves much more like python's ABCs
are expected to, and allows extendability by other abstract classes.